requirements.txt
anyio==3.5.0
asgiref==3.5.0
bcrypt==3.2.0
cachelib==0.1.1
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
click==7.1.2
colorama==0.4.4
Django==3.2.2
djangorestframework==3.13.1
fastapi==0.75.2
fastapi-sessions==0.3.2
Flask==1.1.2
Flask-Session==0.3.2
greenlet==1.1.0
gunicorn==20.1.0
h11==0.13.0
httptools==0.4.0
idna==2.10
itsdangerous==2.1.2
Jinja2==2.11.3
markdown2==2.4.0
MarkupSafe==1.1.1
mysqlclient==2.1.0
passlib==1.7.4
psycopg2-binary==2.8.6
pycparser==2.21
pydantic==1.9.0
python-dotenv==0.17.1
python-multipart==0.0.5
pytz==2021.1
PyYAML==6.0
requests==2.25.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.15
sqlparse==0.4.1
starlette==0.17.1
typing_extensions==4.2.0
urllib3==1.26.4
uvicorn==0.17.6
watchgod==0.8.2
websockets==10.3
Werkzeug==1.0.1
